Added DeleteAvatarService

+++ backend/src/app/controllers/DeliverymanController.js
@@ -2,6 +2,10 @@ import { Op } from 'sequelize';
 import Deliveryman from '../models/Deliveryman';
 import File from '../models/File';
 
+import DeleteAvatarService from '../services/DeleteAvatarService';
+
+import Cache from '../../lib/Cache';
+
 class DeliverymanController {
   async index(req, res) {
     const { page = 1, limit = 10, q, all, order } = req.query;
@@ -43,6 +47,10 @@ class DeliverymanController {
   async show(req, res) {
     const { id } = req.params;
 
+    const cachedKey = `deliveryman:${id}`;
+    const cached = await Cache.get(cachedKey);
+    if (cached) return res.json(cached);
+
     if (!id) return res.status(400).json({ error: 'ID param invald' });
 
     const deliveryman = await Deliveryman.findByPk(id, {
@@ -50,6 +58,8 @@ class DeliverymanController {
         { model: File, as: 'avatar', attributes: ['id', 'path', 'url'] },
       ],
     });
+
+    Cache.set(cachedKey, deliveryman);
 
     return res.json(deliveryman);
   }
@@ -66,6 +76,9 @@ class DeliverymanController {
     if (!deliveryman)
       return res.status(400).json({ error: 'Deliveryman not found' });
 
+    if (avatar_id && avatar_id !== deliveryman.avatar_id)
+      await DeleteAvatarService.run({ avatarId: deliveryman.avatar_id });
+
     const existDeliveryman = await Deliveryman.findOne({ where: { email } });
 
     if (existDeliveryman && existDeliveryman.email !== deliveryman.email)
@@ -73,13 +86,25 @@ class DeliverymanController {
 
     await deliveryman.update({ name, email, avatar_id });
 
+    const cachedKey = `deliveryman:${id}`;
+    await Cache.invalidate(cachedKey);
+
     return res.json(deliveryman);
   }
 
   async delete(req, res) {
     const { id } = req.params;
 
-    await Deliveryman.destroy({ where: { id } });
+    const deliveryman = await Deliveryman.findByPk(id);
+
+    if (deliveryman) {
+      await DeleteAvatarService.run({ avatarId: deliveryman.avatar_id });
+
+      await Deliveryman.destroy({ where: { id } });
+    }
+
+    const cachedKey = `deliveryman:${id}`;
+    await Cache.invalidate(cachedKey);
 
     return res.json();
   }
